menu launches card demo!
I worked Adrian's card demo into my main menu demo, so that my menu displays in the right scale (I had to fit it to the tiny viewport that was being used for the card shuffler, so it doesn't look exactly the same, but it will work for the demo). I'm not a hundred percent that the card demo is working perfect tho, I can't see exactly what its up to.