added logout and pass check
added logout button that seems to work good. Also added logic in the create.php to check that the two passwords match before creating, as well as in the readInput script, now the create button is not even clickable until the passwords match as well as the length requirement from before. Also the menu has been put back together so there shouldn't be any null references or anything now. One thing is that when you log in or create an account, it still doesn't exit the menu yet, and the inputfeilds save their info until you restart the game, which I want to change so they clear when you leave the menu.