new images
designed a few new images for the game's UI, with new images for the buttons (trying to go for a ink brush stroke kind of look), and a new main menu  screen (not really in love with the design currently at all, just something I whipped up quick. Text is kind of hard to read on the buttons, and I still need to change the button animations to work differently.