redisgned UI again (new images and theme)
changed a few things about the images I made for the redesign of the UI, and added a new one to contrast the input fields and menu title's text better. Also changed the button animation to be more readabel. Not sold on the red white green and black colour scheme yet (Although just red white and black look pretty slick).