SERVER BUILDS DECK AND tries to deal
finally got a working version of the multiplayer going! the clients each send their colours to the server, and when it receives two valid colours, it builds the corresponding deck, and tries to deal out the appropriate number of cards to each player! just doesn't quite deal the cards out properly yet.