almost got deal working
So currently the host will deal the cards to the right place (client's hand or oppHand according to if they own the card or not), but the client doesn't spawn the cards in the right place yet despite using the same logic. They get the right cards however, which took a surprsing amount of work to get going. Should be able to crack the issue in the next few days.