Recreated SampleScene to fix peertest#2 merge issues
Took cardDealcopy branch's version of sample scene file, replaced multiplayertroubleshoot's version (since it was full of errors), and then copied peertest's sample scene by hand into multiplayertrouble's version by running an editor for each at the same time and connecting all the pieces. Now the cards colour properly (attack or play), the guide messages show up, and the multiplayer progress is now up to date. The camera is supposed to be set up differently I think (I'm not sure exactly how is working in the peer test so I'm going to need Adrian to go over it, but it should be a very easy fix.