fixed card deal issue! added backside of cards, fixed sound issues, new images
so Adrian and I finally got the cards dealing and displaying properly!!!!! Then I finished making the cards appear in the enemy hand as the backside of the cards (so you can't see your opponent's hand). I then fixed the mute toggle in the sound settings menu to use a script and actually turn the music back on properly. I also added the logo for the game that Henry made a bit ago to be the backs of the cards. Lastly I also updated the main menu art to reflect the newer colour change which looks a bit better, but is not perfect.