fixed issue with duplicated combineddeck on host's gamedeck
Had an issue where the host build of the game's sharedvarmanager was duplicating the combined deck resulting in 80 cards being in the host's version with the proper 40 showing up on the client. Despite the populating the combinedDeck list inside commands only (expected to only be run on server and not either client), when the rpc call to start the game triggers the clients to add the combinedDeck to their version of combined deck, the host's version ends up adding it twice. The server must be running on it's clients version of the combinedDeck to populate it, and then the server calls the rpcs to run on the clients, and the same client gets it twice. I've added another list called temp_combinedDeck, and it is now what the server uses as its gamedeck, and sends it to the rpc call instead of using the same deck. Now neither client has anything in the combinedDeck list until that rpc is called, so there is now duplicate, and both receive the same 40 card deck as intended.