Update SharedVarManager.cs to fix dupliate combinedDeck on Host
Fixed bug with host version of game storing 80 cards in its version of the combinedDeck while client will store proper 40. Its a result of the rpc that runs on both clients, and adds the combinedDeck into whatever they have in combinedDeck at the time of the call (and the host already has the 40 cards from the server creating it, while the client doesn't have anything, which results in host's client adding the complete deck into itself --80 cards--and the other client just receiving the proper 40). Fixed it on multiplayertroubleshooting branch first, just bringing it here