updated both scenes to look better and added play screen
So changed the UI of main menu to be sized better for the camera (it was messed up from the merge conflicts I think). I also changed the samplescene layout as well to include a main start screen with a play button and status message to get around players sending their colours twice by pressing the play button twice and getting the server to start the game with their deck duplicated instead of waiting for the other client to submit their colour. Now they can't see anything except the button and the muiltplayer controls until they click deal, then the button dissapears (avoiding multiple clicks), and the status updates to say they have to wait fo rthe other client, and then it all dissapears to show the main game screen once the other client successfully sends their colour. Also switched around some game elements in the scene to fit within the main canvas, as well as not overlap with the multiplayer controls (host and client networking stuff).