Turns work! can't drag opp's cards, and can only play on your turn!
Got a solid chunk of the turn system done. SharedVarManager tracks the turn count, who's turn it is, and the server assigns each playerManager the proper player number (first to send their colour to server is player 1 and gets to got first). The server updates who's turn it is properly, each client can't change the turn if it isn't their turn, they can only drag cards that are theirs. they can only play during their turn, and the clients update which cards are played and where properly, as well as track who's turn it is as well. Need to get draw working again after each turn still, and health isn't on the server quite yet, but almost there!